Refactor Toaster to MVP
